+++ lib/capistrano/configuration.rb
@@ -110,8 +110,25 @@ module Capistrano
       @timestamp ||= Time.now.utc
     end
 
+    def add_filter(filter=nil, &block)
+      if block
+        raise ArgumentError, "Both a block and an object were given" if filter
+
+        filter = Object.new
+        def filter.filter(servers)
+          block.call(servers)
+        end
+      elsif !filter.respond_to? :filter
+        raise TypeError, "Provided custom filter <#{filter.inspect}> does " \
+                         "not have a public 'filter' method"
+      end
+      @custom_filters ||= []
+      @custom_filters << filter
+    end
+
     def setup_filters
-      @filters = cmdline_filters.clone
+      @filters = cmdline_filters
+      @filters += @custom_filters if @custom_filters
       @filters << Filter.new(:role, ENV["ROLES"]) if ENV["ROLES"]
       @filters << Filter.new(:host, ENV["HOSTS"]) if ENV["HOSTS"]
       fh = fetch_for(:filter, {}) || {}