+++ spec/lib/capistrano/configuration_spec.rb
@@ -304,5 +304,54 @@ module Capistrano
         expect(config.dry_run?).to eq(true)
       end
     end
+
+    describe "custom filtering" do
+      it "accepts a custom filter object" do
+        filter = Object.new
+        def filter.filter(servers)
+          servers
+        end
+        config.add_filter(filter)
+      end
+
+      it "accepts a custom filter as a block" do
+        config.add_filter { |servers| servers }
+      end
+
+      it "raises an error if passed a block and an object" do
+        filter = Object.new
+        def filter.filter(servers)
+          servers
+        end
+
+        expect { config.add_filter(filter) { |servers| servers } }.to raise_error(ArgumentError)
+      end
+
+      it "raises an error if the filter lacks a filter method" do
+        filter = Object.new
+        expect { config.add_filter(filter) }.to raise_error(TypeError)
+      end
+
+      it "calls the filter method of a custom filter" do
+        ENV.delete "ROLES"
+        ENV.delete "HOSTS"
+
+        servers = Configuration::Servers.new
+
+        servers.add_host("test1")
+        servers.add_host("test2")
+        servers.add_host("test3")
+
+        filtered_servers = servers.take(2)
+
+        filter = mock("custom filter")
+        filter.expects(:filter)
+              .with { |subset| subset.is_a? Configuration::Servers }
+              .returns(filtered_servers)
+
+        config.add_filter(filter)
+        expect(config.filter(servers)).to eq(filtered_servers)
+      end
+    end
   end
 end